# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: FrameForward API
  version: 0.1.0
  description: Client and server contract for FrameForward.
  license:
    name: Proprietary
    identifier: LicenseRef-FrameForward-Proprietary
servers:
  - url: /
security: []
paths:
...
  /courses:
    get:
      operationId: listCourses
      summary: List cached P0 courses without generating new content.
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Cached course summaries, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/CourseSummary'}}}}}
        '401': {description: Authentication required.}
  /courses/{courseId}:
    get:
      operationId: getCourse
      summary: Open a cached immutable course version.
      security: [{bearerAuth: []}]
      parameters: [{in: path, name: courseId, required: true, schema: {type: string}}]
      responses: {'200': {description: Course and versioned structured chapters, content: {application/json: {schema: {$ref: '#/components/schemas/CourseDetail'}}}}, '404': {description: Course not found}}
  /courses/{courseId}/progress:
    get:
      operationId: getCourseProgress
      summary: Get progress bound to the opened course content version.
      security: [{bearerAuth: []}]
      parameters: [{in: path, name: courseId, required: true, schema: {type: string}}]
      responses: {'200': {description: Version-bound progress, content: {application/json: {schema: {$ref: '#/components/schemas/CourseProgress'}}}}, '401': {description: Authentication required.}}
  /courses/{courseId}/lessons/{lessonId}/assignments:
    post:
      operationId: submitLessonAssignment
      summary: Submit a valid JPEG assignment for lesson-focused AI feedback.
      security: [{bearerAuth: []}]
      parameters: [{in: header, name: Idempotency-Key, required: true, schema: {type: string}}, {in: path, name: courseId, required: true, schema: {type: string}}, {in: path, name: lessonId, required: true, schema: {type: string}}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AssignmentRequest'}}}}
      responses: {'201': {description: Assignment accepted and lesson feedback started, content: {application/json: {schema: {$ref: '#/components/schemas/LessonFeedback'}}}}, '401': {description: Authentication required.}}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  schemas:
    CourseSummary:
      type: object
      required: [id, title, category, contentVersion, lessonCount]
      properties:
        id: {type: string}
        title: {type: string}
        category: {type: string, enum: [BASICS, MIRRORLESS, EQUIPMENT, MODEL]}
        contentVersion: {type: string}
        lessonCount: {type: integer, minimum: 0}
    CourseDetail:
      type: object
      required: [id, title, category, contentVersion, chapters]
      properties:
        id: {type: string}
        title: {type: string}
        category: {type: string, enum: [BASICS, MIRRORLESS, EQUIPMENT, MODEL]}
        contentVersion: {type: string}
        chapters: {type: array, items: {$ref: '#/components/schemas/Chapter'}}
    Chapter:
      type: object
      required: [id, title, sequence, lessons]
      properties:
        id: {type: string}
        title: {type: string}
        sequence: {type: integer, minimum: 1}
        lessons: {type: array, items: {$ref: '#/components/schemas/Lesson'}}
    Lesson:
      type: object
      required: [id, title, objective, content, correctExample, incorrectExample, exercise, assignment]
      properties:
        id: {type: string}
        title: {type: string}
        objective: {type: string}
        content: {type: string}
        correctExample: {type: string}
        incorrectExample: {type: string}
        exercise: {$ref: '#/components/schemas/JudgmentExercise'}
        assignment: {type: string}
    JudgmentExercise:
      type: object
      required: [question, answer, explanation]
      properties:
        question: {type: string}
        answer: {type: boolean}
        explanation: {type: string}
    CourseProgress:
      type: object
      required: [courseId, contentVersion, completedLessons, totalLessons]
      properties: {courseId: {type: string}, contentVersion: {type: string}, completedLessons: {type: integer, minimum: 0}, totalLessons: {type: integer, minimum: 1}}
    AssignmentRequest:
      type: object
      required: [mediaId]
      properties: {mediaId: {type: string, format: uuid}}
    LessonFeedback:
      type: object
      required: [lessonObjective, feedbackTaskId, progress]
      properties: {lessonObjective: {type: string}, feedbackTaskId: {type: string}, progress: {$ref: '#/components/schemas/CourseProgress'}}
...